pyproject.toml
[project]
name = "fastapi-pg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.3",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "cachetools>=7.0.0",
    "cryptography>=46.0.5",
    "fastapi[standard]>=0.128.0",
    "greenlet>=3.3.1",
    "passlib>=1.7.4",
    "pendulum>=3.2.0",
    "psutil>=7.2.1",
    "pydantic-settings>=2.12.0",
    "pydantic>=2.12.5",
    "pyjwt>=2.11.0",
    "sqlalchemy>=2.0.46",
    "ulid-py>=1.1.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.14.13",
    "ty>=0.0.12",
]


[tool.ty.environment]
python-version = "3.13"
